+++ crates/fuel-core/src/query/balance/asset_query.rs
@@ -4,9 +4,9 @@ use crate::{
         CoinQueryContext,
         MessageQueryContext,
     },
-    state::IterDirection,
 };
 use fuel_core_storage::{
+    iter::IterDirection,
     Error as StorageError,
     Result as StorageResult,
 };
