+++ crates/fuel-core/src/state/rocks_db.rs
@@ -305,15 +305,6 @@ impl BatchOperations for RocksDb {
 
 impl TransactableStorage for RocksDb {}
 
-impl From<IterDirection> for rocksdb::Direction {
-    fn from(d: IterDirection) -> Self {
-        match d {
-            IterDirection::Forward => rocksdb::Direction::Forward,
-            IterDirection::Reverse => rocksdb::Direction::Reverse,
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
