+++ crates/storage/src/iter.rs
@@ -29,3 +29,27 @@ where
         }
     }
 }
+
+/// A enum for iterating across the database
+#[derive(Copy, Clone, Debug, PartialOrd, Eq, PartialEq)]
+pub enum IterDirection {
+    /// Iterate forward
+    Forward,
+    /// Iterate backward
+    Reverse,
+}
+
+impl Default for IterDirection {
+    fn default() -> Self {
+        Self::Forward
+    }
+}
+
+impl From<IterDirection> for rocksdb::Direction {
+    fn from(d: IterDirection) -> Self {
+        match d {
+            IterDirection::Forward => rocksdb::Direction::Forward,
+            IterDirection::Reverse => rocksdb::Direction::Reverse,
+        }
+    }
+}
